useSearchAggregationOptions as a separated file
